Make modular to be able to export all of any value from the db
